+++ plugins/dates/src/api/dates.js
@@ -2,6 +2,15 @@ import find from './find/index.js'
 import parseDates from './parse/index.js'
 import toJSON from './toJSON.js'
 
+const quickDate = function (view, str) {
+  let tmp = view.fromText(str)
+  let found = parseDates(tmp, view.opts)[0]
+  if (!found || !found.start || !found.start.d) {
+    return null
+  }
+  return found.start.d
+}
+
 const api = function (View) {
   class Dates extends View {
     constructor(document, pointer, groups, opts = {}) {
@@ -57,18 +66,26 @@ const api = function (View) {
 
     /** return only dates occuring before a given date  */
     isBefore(iso) {
-      let tmp = this.fromText(iso)
-      let found = parseDates(tmp, this.opts)[0]
-      if (!found || !found.start || !found.start.d) {
-        return this.none() //return nothing
-      }
-      let pivot = found.start.d // our given date
+      let pivot = quickDate(this, iso)
       return this.filter(m => {
         let obj = parseDates(m, this.opts)[0] || {}
-        if (!obj.start || !obj.start.d) {
-          return false
-        }
-        return obj.start.d.isBefore(pivot)
+        return obj.start && obj.start.d && obj.start.d.isBefore(pivot)
+      })
+    }
+    /** return only dates occuring after a given date  */
+    isAfter(iso) {
+      let pivot = quickDate(this, iso)
+      return this.filter(m => {
+        let obj = parseDates(m, this.opts)[0] || {}
+        return obj.start && obj.start.d && obj.start.d.isAfter(pivot)
+      })
+    }
+    /** return only dates occuring after a given date  */
+    isSame(unit, iso) {
+      let pivot = quickDate(this, iso)
+      return this.filter(m => {
+        let obj = parseDates(m, this.opts)[0] || {}
+        return obj.start && obj.start.d && obj.start.d.isSame(pivot, unit)
       })
     }
   }